refactor: rename LongStudy to longitudinal_study
